Refactor executeProcess so that it's more readable.
Research the shortcut to toggle flag that was shown to us months ago

Use the toggle to toggle the this.stateChanged flag

No new tests pass, but it looks cleaner and more evident of what it does

Queue

Flip two statements in enqueue so that it would work.

Reset quantumClock to 0 if state changes

Take off var declaration for dequeue. Don't need it.

Edit handleInterrupt so that it comes from correct class.